+++ api/app.js
@@ -1,9 +1,11 @@
 import express from 'express'
 import path from 'path'
 import userRoutes from './routes/userRoutes.js'
+import cookieParser from "cookie-parser";
 
 const app = express()
 
+app.use(cookieParser());
 app.use(express.json())
 app.use(express.static(path.resolve('public')))