+++ api/middleware/cookieAuth.js
@@ -1,10 +1,16 @@
 import jwt from "jsonwebtoken";
 
 export function cookieAuth(req, res, next){
-  try {
-    console.log("middleware, ebaaaa")
+  try {    
+    const refreshToken = req.cookies.refreshToken
+    if (!refreshToken) return res.sendStatus(401);
+
+    jwt.verify(refreshToken, process.env.REFRESH_TOKEN, (err, user) => {
+      if (err) return res.sendStatus(403);
+    });
+      
     next()
   } catch (error) {
-    console.log("moiou, ó:", error)
+    res.status(500).send({message: "Erro interno do servidor:", error})
   }
 }