+++ api/routes/userRoutes.js
@@ -8,7 +8,8 @@ userRoutes.post('/', UserController.create)
 userRoutes.post('/login', cookieAuth, UserController.login)
 userRoutes.get('/', cookieAuth, UserController.list)
 userRoutes.get('/:id', cookieAuth, UserController.getOne)
-userRoutes.put('/:id', cookieAuth, UserController.update)
-userRoutes.delete('/:id', cookieAuth, UserController.delete)
+userRoutes.post('/atualizar', cookieAuth, UserController.updateMultiple)
+userRoutes.post('/deletar', cookieAuth, UserController.deleteMultiple)
+// userRoutes.delete('/:id', cookieAuth, UserController.delete)
 
 export default userRoutes